Physics and Collision Detection
I created a class PhysicsEngine which takes in the screen width and
height as its constructor parameters. This class is responsible for
iterating through each object's forces or velocities and applying them
accordingly as well as updating those forces such that they attenuate
over time.

The PhysicsEngine class also has a function triggerImpact that takes 2
objects and calculates their respective velocities and determines the
resulting impact force to be applied to both of them. In this function
it also determines the space needed to be translated between the objects
on collision such that after all forces are applied they are no longer
overlapping which would result in infinite collisions detected

Changed the Force class to Velocity class due to request. This class is
just a vector, it has an x y and z value which are its strengths in
those directions. It also has a duration which at the moment is measured
in frames and will later be changed to be measured in seconds as well as
a timeLeft counter that indicates how much time this force or velocity
has left before it is no longer in effect and a strength variable which
is just timeLeft / duration

the way it is set up, the physics engine can deal with object given that
those objects have a vector of velocities to check against, a mass, and
a rectangle representing its top down view of the area it takes up (ie
the bottom face of its bounding box)